pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "brainbuilding"
version = "0.1.0"
description = "A package for brain building and analysis"
authors = [
    {name = "Nikita Smirnov", email = "neket2000@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

# Minimal runtime deps for realtime EEG service
dependencies = [
    "numpy>=2.1.3",
    "scipy>=1.12.0",
    "scikit-learn>=1.4.0",
    "pyriemann>=0.7",
    "pylsl==1.16.2",
    "pydantic>=2.0.0",
    "PyYAML>=6.0.1",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
# Training-only extras (not needed for realtime service)
train = [
    "mne>=1.8.0",
    "pyneurostim @ git+https://github.com/VladimirR46/pyneurostim.git@b022759f96466304ad600c0f48fb5397bd6fcd17",
    "pandas>=2.2.3",
    "matplotlib>=3.9.2",
    "seaborn>=0.13.0",
    "tqdm>=4.66.0",
    "h5py>=3.10.0",
    "pyxdf>=1.16.5",
]

[tool.setuptools]
packages = ["brainbuilding"] 

[project.scripts]
brainbuilding-realtime = "brainbuilding.service.cli:main"

[tool.ruff]
line-length = 79
extend-exclude = [
    "build",
    "data",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true